// TransformationPlayground/playground_frontend/src/generated-api-client/api.ts
export const SourceVersionEnum = {
    _68: 'Elasticsearch 6.8',
    _710: 'Elasticsearch 7.10'
} as const;

export type SourceVersionEnum = typeof SourceVersionEnum[keyof typeof SourceVersionEnum];


/**
 * * `OpenSearch 2.17` - OS_2_17
 * @export
 * @enum {string}
 */

export const TargetVersionEnum = {
    OpenSearch217: 'OpenSearch 2.17'
} as const;

export type TargetVersionEnum = typeof TargetVersionEnum[keyof typeof TargetVersionEnum];


/**
 * * `Python` - PYTHON
 * @export
 * @enum {string}
 */

export const TransformLanguageEnum = {
    Python: 'Python'
} as const;

export type TransformLanguageEnum = typeof TransformLanguageEnum[keyof typeof TransformLanguageEnum];


/**
 * 
 * @export
 * @interface TransformsIndexCreateRequest
 */
export interface TransformsIndexCreateRequest {
    /**
     * 
     * @type {SourceVersionEnum}
     * @memberof TransformsIndexCreateRequest
     */
    'source_version': SourceVersionEnum;
    /**
     * 
     * @type {TargetVersionEnum}
     * @memberof TransformsIndexCreateRequest
     */
    'target_version': TargetVersionEnum;
    /**
     * 
     * @type {TransformLanguageEnum}
     * @memberof TransformsIndexCreateRequest
     */
    'transform_language': TransformLanguageEnum;
    /**
     * 
     * @type {TransformsIndexCreateRequestInputShape}
     * @memberof TransformsIndexCreateRequest
     */
    'input_shape': TransformsIndexCreateRequestInputShape;
    /**
     * 
     * @type {string}
     * @memberof TransformsIndexCreateRequest
     */
    'transform_logic'?: string;
    /**
     * 
     * @type {string}
     * @memberof TransformsIndexCreateRequest
     */
    'user_guidance'?: string;
    /**
     * 
     * @type {string}
     * @memberof TransformsIndexCreateRequest
     */
    'test_target_url'?: string;
}


/**
 * 
 * @export
 * @interface TransformsIndexCreateRequestInputShape
 */
export interface TransformsIndexCreateRequestInputShape {
    /**
     * 
     * @type {string}
     * @memberof TransformsIndexCreateRequestInputShape
     */
    'index_name': string;
    /**
     * 
     * @type {object}
     * @memberof TransformsIndexCreateRequestInputShape
     */
    'index_json': object;
}
// ...
export interface TransformsIndexTestRequest {
    /**
     * 
     * @type {TransformLanguageEnum}
     * @memberof TransformsIndexTestRequest
     */
    'transform_language': TransformLanguageEnum;
    /**
     * 
     * @type {SourceVersionEnum}
     * @memberof TransformsIndexTestRequest
     */
    'source_version': SourceVersionEnum;
    /**
     * 
     * @type {TargetVersionEnum}
     * @memberof TransformsIndexTestRequest
     */
    'target_version': TargetVersionEnum;
    /**
     * 
     * @type {TransformsIndexCreateRequestInputShape}
     * @memberof TransformsIndexTestRequest
     */
    'input_shape': TransformsIndexCreateRequestInputShape;
    /**
     * 
     * @type {string}
     * @memberof TransformsIndexTestRequest
     */
    'transform_logic': string;
    /**
     * 
     * @type {string}
     * @memberof TransformsIndexTestRequest
     */
    'test_target_url'?: string;
}
// ...
export class TransformsApi extends BaseAPI {
    /**
     * 
     * @param {TransformsIndexCreateRequest} transformsIndexCreateRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof TransformsApi
     */
    public transformsIndexCreateCreate(transformsIndexCreateRequest: TransformsIndexCreateRequest, options?: RawAxiosRequestConfig) {
        return TransformsApiFp(this.configuration).transformsIndexCreateCreate(transformsIndexCreateRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {TransformsIndexTestRequest} transformsIndexTestRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof TransformsApi
     */
    public transformsIndexTestCreate(transformsIndexTestRequest: TransformsIndexTestRequest, options?: RawAxiosRequestConfig) {
        return TransformsApiFp(this.configuration).transformsIndexTestCreate(transformsIndexTestRequest, options).then((request) => request(this.axios, this.basePath));
    }
}